refactor: Extract generic interface for future multis
